rename constants.py into dku_constants.py
