+++ python-lib/gluonts_forecasts/gluon_dataset.py
@@ -1,5 +1,5 @@
 from gluonts.dataset.common import ListDataset
-from constants import TIMESERIES_KEYS
+from dku_constants import TIMESERIES_KEYS
 import numpy as np
