+++ python-lib/gluonts_forecasts/utils.py
@@ -4,7 +4,7 @@ from pandas.tseries.frequencies import to_offset
 import numpy as np
 from functools import reduce
 import copy
-from constants import TIMESERIES_KEYS, ROW_ORIGIN, CUSTOMISABLE_FREQUENCIES_OFFSETS, GPU_CONFIGURATION
+from dku_constants import TIMESERIES_KEYS, ROW_ORIGIN, CUSTOMISABLE_FREQUENCIES_OFFSETS, GPU_CONFIGURATION
 from timeseries_preparation.preparation import TimeseriesPreparator
