+++ tests/python/unit/test_missing_values.py
@@ -3,7 +3,7 @@ from gluonts_forecasts.model import Model
 from timeseries_preparation.preparation import TimeseriesPreparator
 from gluonts_forecasts.gluon_dataset import GluonDataset
 from gluonts_forecasts.model_handler import MODEL_DESCRIPTORS, LABEL, list_available_models
-from constants import METRICS_DATASET, EVALUATION_METRICS_DESCRIPTIONS, TIMESERIES_KEYS
+from dku_constants import METRICS_DATASET, EVALUATION_METRICS_DESCRIPTIONS, TIMESERIES_KEYS
 from datetime import datetime
 from pandas.api.types import is_datetime64_ns_dtype
 import pandas as pd
